requirements.txt
streamlit==1.30.0
streamlit-option-menu==0.3.12
pandas==1.5.3
numpy==1.24.2
tensorflow==2.15.0
scikit-learn==1.5.1
plotly==5.22.0
seaborn==0.13.2
matplotlib==3.9.0
joblib==1.4.2
imbalanced-learn==0.12.3
Pillow==9.4.0